Fix content insets calculation error when taking task snapshot.
Previously, we are using the task bound as frame to calculate system bar insets when taking task snapshot so the system bar insets contain the status bar and navigation bar insets. However, for some activity the window frame has already excluded the status bar area which is filled by letterbox. Then the letter box insets is added to the content insets and so the status bar area is counted twice, causing the task's snapshot to be smaller than its actual window size. As the consequence, the snapshot starting window cannot cover the original activity, giving the users a impression that the window has been 'pulled downward'.
Bug: 209930559

Signed-off-by: mashixing <mashixing@xiaomi.corp-partner.google.com>
Change-Id: I356ebd13a58e5411d641435e01df7021f5673d9b
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>